dummy_repo/requirements.txt
flask
requests
mcp-server
@modelcontextprotocol/sdk
